Migrated to postgres and few changes to project structure

+++ src/app/api/user/[userid]/transactions/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import prisma from '@/prisma/client';
+import prisma from '@/src/lib/prisma';
 
 export async function GET(request: NextRequest, { params }: { params: { userid: string } }) {
   const { userid } = params;
@@ -11,6 +11,7 @@ export async function GET(request: NextRequest, { params }: { params: { userid: 
 
     return NextResponse.json(transactions);
   } catch (error) {
+    console.error('Error retrieving transactions:', error);
     return NextResponse.json({ message: 'Error retrieving transactions', error }, { status: 500 });
   }
 }
@@ -23,16 +24,19 @@ export async function POST(request: NextRequest, { params }: { params: { userid:
     const newTransaction = await prisma.transaction.create({
       data: {
         userId: userid,
-        coinName: body.coinName,
-        coinAmount: body.coinAmount,
-        coinPrice: body.coinPrice,
-        coinSymbol: body.coinSymbol,
-        transactionDate: body.transactionDate
+        name: body.name,
+        amount: body.amount,
+        price: body.price,
+        symbol: body.symbol,
+        value: body.amount * body.price,
+        date: new Date(body.date),
+        type: body.type,
       }
     });
 
     return NextResponse.json(newTransaction);
   } catch (error) {
+    console.error('Error creating transaction:', error);
     return NextResponse.json({ message: 'Error creating transaction', error }, { status: 500 });
   }
 }
@@ -43,15 +47,26 @@ export async function DELETE(request: NextRequest, { params }: { params: { useri
   const transactionId = body.id;
 
   try {
-    await prisma.transaction.delete({
+    const transaction = await prisma.transaction.findFirst({
       where: {
         id: transactionId,
-        userId: userid 
+        userId: userid
+      }
+    });
+
+    if (!transaction) {
+      return NextResponse.json({ message: 'Transaction not found' }, { status: 404 });
+    }
+
+    await prisma.transaction.delete({
+      where: {
+        id: transactionId
       }
     });
 
     return NextResponse.json({ message: 'Transaction deleted successfully' });
   } catch (error) {
+    console.error('Error deleting transaction:', error);
     return NextResponse.json({ message: 'Error deleting transaction', error }, { status: 500 });
   }
 }
@@ -62,13 +77,10 @@ export async function PUT(request: NextRequest, { params }: { params: { userid: 
   const transactionId = body.id;
 
   try {
-    const existingTransaction = await prisma.transaction.findUnique({
+    const existingTransaction = await prisma.transaction.findFirst({
       where: {
         id: transactionId,
-      },
-      select: {
-        coinName: true,
-        coinSymbol: true,
+        userId: userid
       }
     });
 
@@ -79,20 +91,21 @@ export async function PUT(request: NextRequest, { params }: { params: { userid: 
     const updatedTransaction = await prisma.transaction.update({
       where: {
         id: transactionId,
-        userId: userid,
       },
       data: {
-        coinAmount: body.coinAmount,
-        coinPrice: body.coinPrice,
-        transactionDate: body.transactionDate,
-        coinName: existingTransaction.coinName,
-        coinSymbol: existingTransaction.coinSymbol,
+        amount: body.amount,
+        price: body.price,
+        date: new Date(body.date),
+        type: body.type,
+        name: existingTransaction.name,  // Retain original name
+        symbol: existingTransaction.symbol, // Retain original symbol
+        value: body.amount * body.price, // Recalculate the value
       }
     });
 
     return NextResponse.json(updatedTransaction);
   } catch (error) {
+    console.error('Error updating transaction:', error);
     return NextResponse.json({ message: 'Error updating transaction', error }, { status: 500 });
   }
 }
-
